+++ skyfield/tests/test_timelib.py
@@ -9,7 +9,7 @@ from datetime import datetime
 one_second = 1.0 / DAY_S
 epsilon = one_second * 42.0e-6  # 20.1e-6 is theoretical best precision
 
-time_parameter = ['tai', 'tt', 'tdb']
+time_parameter = ['tai', 'tt', 'tdb', 'ut1']
 time_value = [(1973, 1, 18, 1, 35, 37.5), 2441700.56640625]
 
 def ts():
@@ -65,6 +65,21 @@ def test_building_time_from_list_of_utc_datetimes(ts):
     assert (t.tai == [
         2442046.5, 2442047.5, 2442048.5, 2442049.5, 2442050.5, 2442051.5,
         ]).all()
+
+def test_converting_ut1_to_tt(ts):
+    ten_thousand_years = 365 * 10000
+
+    jd = api.T0 - ten_thousand_years
+    t = ts.ut1(jd=jd)
+    del t.ut1                   # force re-computation of UT1
+    print(jd - t.ut1)
+    assert abs(jd - t.ut1) < 1e-10
+
+    jd = api.T0 + ten_thousand_years
+    t = ts.ut1(jd=jd)
+    del t.ut1                   # force re-computation of UT1
+    print(jd - t.ut1)
+    assert abs(jd - t.ut1) < 1e-10
 
 def test_indexing_time(ts):
     t = ts.utc(1974, 10, range(1, 6))